upload 7/18 app project update
Adding more code to the app project. Mostly work on the "pins" page. Setting up check boxes for each pin, writing scripts to create arrays for each series of pins the user has/needs. Creating a master list for the user and a progress bar on total number of unique pins they have.